Finish documentation for entity.py
