+++ test/scanner/array.py
@@ -3,7 +3,7 @@
 """
 
 from test.scanner import JSPECTestScanner
-from jspec.component import (
+from jspec.entity import (
     JSPEC, 
     JSPECArray,
     JSPECInt,
