+++ test/scanner/placeholder.py
@@ -2,7 +2,7 @@
 """
 
 from test.scanner import JSPECTestScanner
-from jspec.component import (
+from jspec.entity import (
     JSPEC,
     JSPECObjectPlaceholder,
     JSPECArrayPlaceholder,
